Add toggleCase command
